requirements.txt
six
urllib3
certifi
requests
python-dotenv